feat:상단바 유저 정보 추가 #110

+++ src/main/java/com/tickerbell/jujuclub/auth/dto/MemberDTO.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MemberDTO {
-    private Long userSeq;
+    private Integer userSeq;
     private String userId;
     private String userPw;
     private String userName;
